* master:
  Implement mining control features (step size, max fan speed, max chip temp, optimization mode)
  replair siahash difficulty
  fix dcr difficulty mistake
  fix: Typo "Privision" -> "Provision"

# Conflicts:
#	src/apiserver/src/handlers/Handlers.cpp
#	src/cgminer/cgminer.c
#	src/cgminer/driver-obelisk.c
#	src/cgminer/miner.h
#	src/cgminer/obelisk/CSS_SC1Defines.h
#	src/cgminer/obelisk/Ob1FanCtrl.c
#	src/webclient/src/modules/Main/reducer.ts
#	src/webclient/src/modules/Main/types.ts
#	src/webclient/src/panels/MiningPanel.tsx
#	src/webclient/src/panels/SystemPanel.tsx

+++ src/cgminer/obelisk/siahash/siaverify.c
@@ -49,7 +49,7 @@ bool siaHeaderMeetsProvidedTarget(uint8_t header[80], uint8_t target[32]) {
 // which is different from a Sia difficulty.
 bool siaHeaderMeetsProvidedDifficulty(uint8_t header[80], double difficulty) {
 	uint64_t baseDifficulty = 0xffff000000000000;
-	uint64_t adjustedDifficulty = baseDifficulty/difficulty;
+	uint64_t adjustedDifficulty = baseDifficulty/(uint64_t)difficulty;
 	uint8_t target[32] = {0x00, 0x00, 0x00, 0x00, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff};
 	int i = 0;
 	for( i = 0; i < 8; i++) {
@@ -66,7 +66,7 @@ bool siaHeaderMeetsProvidedDifficulty(uint8_t header[80], double difficulty) {
 int siaHeaderMeetsChipTargetAndPoolDifficulty(uint8_t header[80], uint8_t chipTarget[32], double poolDifficulty) {
 	// Calculate the target based on the pool difficulty.
 	uint64_t baseDifficulty = 0xffff000000000000;
-	uint64_t adjustedDifficulty = baseDifficulty/poolDifficulty;
+	uint64_t adjustedDifficulty = baseDifficulty/(uint64_t)poolDifficulty;
 	uint8_t poolTarget[32] = {0x00, 0x00, 0x00, 0x00, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff};
 	int i = 0;
 	for( i = 0; i < 8; i++) {
